+++ lib/paper_button_base.dart
@@ -6,15 +6,14 @@ library paper_elements.paper_button_base;
 
 import 'dart:html';
 import 'dart:js' show JsArray, JsObject;
-import 'package:web_components/custom_element_proxy.dart';
-import 'package:web_components/html_import_annotation.dart';
-import 'package:custom_element_apigen/src/common.dart' show PolymerProxyMixin, DomProxyMixin;
+import 'package:web_components/web_components.dart';
+import 'package:polymer_interop/polymer_interop.dart';
 import 'package:core_elements/core_focusable.dart';
 import 'paper_ripple.dart';
 
 /// `paper-button-base` is the base class for button-like elements with ripple and optional shadow.
 @CustomElementProxy('paper-button-base')
-class PaperButtonBase extends HtmlElement with DomProxyMixin, PolymerProxyMixin, CoreFocusable {
+class PaperButtonBase extends HtmlElement with CustomElementProxyMixin, PolymerProxyMixin, CoreFocusable {
   PaperButtonBase.created() : super.created();
   factory PaperButtonBase() => new Element.tag('paper-button-base');
 }