+++ lib/paper_char_counter.dart
@@ -6,9 +6,8 @@ library paper_elements.paper_char_counter;
 
 import 'dart:html';
 import 'dart:js' show JsArray, JsObject;
-import 'package:web_components/custom_element_proxy.dart';
-import 'package:web_components/html_import_annotation.dart';
-import 'package:custom_element_apigen/src/common.dart' show PolymerProxyMixin, DomProxyMixin;
+import 'package:web_components/web_components.dart';
+import 'package:polymer_interop/polymer_interop.dart';
 import 'package:core_elements/core_style.dart';
 
 /// Material Design: <a href="http://www.google.com/design/spec/components/text-fields.html#text-fields-character-counter">Character counter</a>
@@ -27,7 +26,7 @@ import 'package:core_elements/core_style.dart';
 ///
 /// `paper-char-counter` uses `paper-input-decorator`'s error `core-style` for global theming.
 @CustomElementProxy('paper-char-counter')
-class PaperCharCounter extends HtmlElement with DomProxyMixin, PolymerProxyMixin {
+class PaperCharCounter extends HtmlElement with CustomElementProxyMixin, PolymerProxyMixin {
   PaperCharCounter.created() : super.created();
   factory PaperCharCounter() => new Element.tag('paper-char-counter');