+++ lib/paper_radio_button.dart
@@ -6,9 +6,8 @@ library paper_elements.paper_radio_button;
 
 import 'dart:html';
 import 'dart:js' show JsArray, JsObject;
-import 'package:web_components/custom_element_proxy.dart';
-import 'package:web_components/html_import_annotation.dart';
-import 'package:custom_element_apigen/src/common.dart' show PolymerProxyMixin, DomProxyMixin;
+import 'package:web_components/web_components.dart';
+import 'package:polymer_interop/polymer_interop.dart';
 import 'paper_ripple.dart';
 import 'package:core_elements/core_a11y_keys.dart';
 
@@ -53,7 +52,7 @@ import 'package:core_elements/core_a11y_keys.dart';
 ///       border-color: #b5b5b5;
 ///     }
 @CustomElementProxy('paper-radio-button')
-class PaperRadioButton extends HtmlElement with DomProxyMixin, PolymerProxyMixin {
+class PaperRadioButton extends HtmlElement with CustomElementProxyMixin, PolymerProxyMixin {
   PaperRadioButton.created() : super.created();
   factory PaperRadioButton() => new Element.tag('paper-radio-button');